feat: add response annotations and null checks to ProdutoController

+++ ApiFuncional/Controllers/ProdutosController.cs
@@ -3,6 +3,7 @@ using ApiFuncional.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 
 namespace ApiFuncional.Controllers
 {
@@ -18,20 +19,47 @@ namespace ApiFuncional.Controllers
         }
 
         [HttpGet]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<IEnumerable<Produto>>> GetProdutos()
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            } 
             return await _context.Produtos.ToListAsync();
         }
 
         [HttpGet("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> GetProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();
+            }
+
             return produto;
         }
         [HttpPost]
+        [ProducesResponseType(StatusCodes.Status201Created)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> PostProduto(Produto produto)
         {
+            if (_context.Produtos == null)
+            {
+                return Problem("Errro ao criar um produto!");
+            }
             if (!ModelState.IsValid)
             {
                 return ValidationProblem(new ValidationProblemDetails(ModelState)
@@ -45,26 +73,61 @@ namespace ApiFuncional.Controllers
             return CreatedAtAction(nameof(GetProduto), new { id = produto.Id }, produto);
         }
         [HttpPut("{id:int}")]
-        public async Task<ActionResult<Produto>> PutProduto(int id, Produto produto)
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
+        public async Task<ActionResult> PutProduto(int id, Produto produto)
         {
             if (id != produto.Id) return BadRequest();
 
             if (!ModelState.IsValid) return ValidationProblem(ModelState);
 
-            _context.Produtos.Update(produto);
-            await _context.SaveChangesAsync();
-
+            _context.Entry(produto).State = EntityState.Modified;
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                if(!ProdutoExists(id))
+                {
+                    return NotFound();
+                }
+                else 
+                {
+                    throw;
+                }
+            }
             return NoContent();
         }
         [HttpDelete("{id:int}")]
+        [ProducesResponseType(StatusCodes.Status204NoContent)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesDefaultResponseType]
         public async Task<ActionResult<Produto>> DeleteProduto(int id)
         {
+            if (_context.Produtos == null)
+            {
+                return NotFound();
+            }
+            
             var produto = await _context.Produtos.FindAsync(id);
+
+            if (produto == null)
+            {
+                return NotFound();
+            }
 
             _context.Produtos.Remove(produto);
             await _context.SaveChangesAsync();
 
             return NoContent();
+        }
+
+        private bool ProdutoExists(int id)
+        {
+            return (_context.Produtos?.Any(e => e.Id == id)).GetValueOrDefault();
         }
     }
 }
